Reworks legacy code
