+++ packages/server/src/models-abstract/measurement-model.js
@@ -129,80 +129,6 @@ class MeasurementModel extends ModelPostgres {
       GROUP BY "Measurements"."measurableId", "Measurements"."agentId"
     )`;
   }
-
-  /**
-   * @public
-   * @todo: use getConnection
-   * @deprecated
-   * @param {Layers.AbstractModelsLayer.filter} [filter]
-   * @param {Models.ObjectID} [filter.measurableId]
-   * @param {Models.ObjectID} [filter.agentId]
-   * @param {string[]} [filter.competitorType]
-   * @param {Models.ObjectID} [filter.notTaggedByAgent]
-   * @param {Layers.AbstractModelsLayer.pagination} [pagination]
-   * @param {Layers.AbstractModelsLayer.restrictions} [restrictions]
-   * @param {Layers.AbstractModelsLayer.options} [_options]
-   * @return {Promise<{data: Models.Measurement[], total: number}>}
-   */
-  async getAll(filter = {}, pagination = {}, restrictions = {}, _options = {}) {
-    if ('inspect' in filter) filter.inspect();
-    if ('inspect' in pagination) pagination.inspect();
-    if ('inspect' in restrictions) restrictions.inspect();
-
-    const { where, include, spacedLimit } = this.makeFilter(filter);
-    this.applyRestrictions(where, restrictions);
-
-    /** @type {number} */
-    const total = await this.model.count({ where, include });
-    const edgePagination = this._getPagination(pagination, total);
-
-    const cond = {
-      limit: edgePagination.limit,
-      offset: edgePagination.offset,
-      order: [['relevantAt', 'DESC']],
-      where,
-      include,
-    };
-
-    /** @type {Models.Measurement[]} */
-    let data = await this.model.findAll(cond);
-    data = this._setIndexes(data, edgePagination);
-    data.total = total;
-
-    // tricky
-    if (spacedLimit) data = splitBy(data, spacedLimit);
-
-    return new ResponseAll(data, total);
-  }
-
-  /**
-   * @protected
-   * @param {object} filter
-   * @return {{include: Array, spacedLimit: *, where: {}}}
-   */
-  makeFilter(filter = {}) {
-    const where = { [this.and]: [] };
-    const include = [];
-
-    const startDate = _.get(filter, 'findInDateRange.startDate');
-    const endDate = _.get(filter, 'findInDateRange.endDate');
-    const spacedLimit = _.get(filter, 'findInDateRange.spacedLimit');
-
-    if (filter.measurableId) where.measurableId = filter.measurableId;
-    if (filter.agentId) where.agentId = filter.agentId;
-    if (filter.competitorType) where.competitorType = {
-      [this.in]: filter.competitorType,
-    };
-    if (startDate) where[this.and].push({
-      createdAt: { [this.gte]: startDate },
-    });
-    if (endDate) where[this.and].push({ createdAt: { [this.lte]: endDate } });
-    if (filter.notTaggedByAgent) where.id = {
-      [this.notIn]: this._taggedMeasurementsLiteral(filter.notTaggedByAgent),
-    };
-
-    return { where, include, spacedLimit };
-  }
 }
 
 module.exports = {
